rename: datshboard type 파일명을 dashboard로 변경
